Clean up modal styles
