--- src/components/Modals/DownloadPymolScriptModal.styles.js
@@ -1,1 +1,0 @@
-// import styled from 'styled-components';
